refactor(Footer): move agent and client to the same root folder
