+++ packages/css/src/Layout/Footer/Footer.client.scss
@@ -1,4 +1,4 @@
-@use "../common/common.client.scss" as common;
+@use "../../common/common.client.scss" as common;
 
 .af-footer {
   z-index: 100;
