+++ packages/css/src/agent.scss
@@ -11,7 +11,7 @@
 @use "./Form/Textarea/Textarea.agent.scss";
 @use "./Modal/Modal.agent.scss";
 @use "./Title/Title.agent.scss";
-@use "./Layout/Footer/Footer.scss";
+@use "./Layout/Footer/Footer.agent.scss";
 @use "./Badge/Badge.scss";
 @use "./Tabs/Tabs.agent.scss";
 @use "./Alert/Alert.agent.scss";
