+++ packages/css/src/client.scss
@@ -1,6 +1,6 @@
 @use "./Button/Button.client";
 @use "./Header/Header.client";
-@use "./Footer/Footer.client";
+@use "./Layout/Footer/Footer.client.scss";
 @use "./Stepper/Stepper.client";
 @use "./Form/Radio/Client/Radio.scss";
 @use "./Tabs/Tabs.client.scss";
