+++ packages/react/src/Layout/Footer/Client/Footer.mdx
@@ -1,5 +1,5 @@
 import { Canvas, Controls, Meta } from "@storybook/addon-docs";
-import * as FooterStories from "./Footer.client.stories";
+import * as FooterStories from "./Footer.stories";
 
 <Meta of={FooterStories} name="Footer" />
