+++ packages/react/src/client.ts
@@ -7,7 +7,6 @@ export {
   ButtonClient as Button,
   Variants as ButtonVariants,
 } from "./Button/Button.client";
-export { Footer } from "./Footer/Footer.client";
 export {
   Checkbox,
   CheckboxInput,
@@ -16,12 +15,13 @@ export {
 } from "./Form/Checkbox/Client";
 export { Radio, RadioInput, RadioItem, RadioModes } from "./Form/Radio/Client";
 export { Header } from "./Header/Header.client";
+export { Footer } from "./Layout/Footer/Client/Footer";
+export { Link } from "./Link/Link.client";
 export { Modal } from "./ModalClient/Modal.client";
 export type { ButtonAction as ModalButtonAction } from "./ModalClient/Modal.client";
+export { Pagination } from "./Pagination/Pagination.client";
 export { Stepper } from "./Stepper/Stepper.client";
 export {
   TabsClient as Tabs,
   Direction as TabsDirection,
 } from "./Tabs/Client/Tabs.client";
-export { Link } from "./Link/Link.client";
-export { Pagination } from "./Pagination/Pagination.client";
